--- Assets/Scripts/BombRangUpdate.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 9bef5e4f7aaf69f4790234b8434215d2
-timeCreated: 1524052709
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
