+++ Assets/Scripts/Effect.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 6f76aa4c3f3ea0c40b664c505911866c
+folderAsset: yes
+timeCreated: 1524655269
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
