+++ Assets/Scripts/Effect/BombRangUpdate.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class BombRangeUpdate : MonoBehaviour
+{
+    
+    
+    // Use this for initialization
+    void Start()
+    {
+
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+
+    }
+
+    
+}
