+++ Assets/Scripts/Effect/BombUp.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c2f639abf869b3540af98698ed88195b
+timeCreated: 1524488157
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
