--- Assets/Scripts/SpeedUp.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: afcfed6499a351d4db4819731dff2f4f
-timeCreated: 1524486284
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
