+++ Assets/Sound Effects.meta
@@ -1,9 +1,10 @@
 fileFormatVersion: 2
-guid: 1df2e42271f25aa4685d4c64ee9edddd
+guid: 7e2ed1e32a654e04797676f5c1cc8d7e
 folderAsset: yes
-timeCreated: 1454535662
+timeCreated: 1524656661
 licenseType: Free
 DefaultImporter:
+  externalObjects: {}
   userData: 
   assetBundleName: 
   assetBundleVariant:
