+++ Assets/Sound Effects/BombCharge.mp3.meta
@@ -3,6 +3,7 @@ guid: b49377a2a390b5a4b8f5d8529226afa0
 timeCreated: 1454535663
 licenseType: Free
 AudioImporter:
+  externalObjects: {}
   serializedVersion: 6
   defaultSettings:
     loadType: 0
@@ -16,6 +17,7 @@ AudioImporter:
   normalize: 1
   preloadAudioData: 1
   loadInBackground: 0
+  ambisonic: 0
   3D: 1
   userData: 
   assetBundleName:
