+++ BomberMan.csproj
@@ -242,8 +242,9 @@
     <Compile Include="Assets\Scripts\Bomb\BeBomb.cs" />
     <Compile Include="Assets\Scripts\Bomb\CreatExplore.cs" />
     <Compile Include="Assets\Scripts\Bomb\ExitTrigger.cs" />
-    <Compile Include="Assets\Scripts\BombRangUpdate.cs" />
-    <Compile Include="Assets\Scripts\BombUp.cs" />
+    <Compile Include="Assets\Scripts\Effect\BombRangUpdate.cs" />
+    <Compile Include="Assets\Scripts\Effect\BombUp.cs" />
+    <Compile Include="Assets\Scripts\Effect\SpeedUp.cs" />
     <Compile Include="Assets\Scripts\Explore\ClearExplotion.cs" />
     <Compile Include="Assets\Scripts\GameManager.cs" />
     <Compile Include="Assets\Scripts\Player\Player\IMove\IMovement.cs" />
@@ -251,7 +252,6 @@
     <Compile Include="Assets\Scripts\Player\Player\Player2.cs" />
     <Compile Include="Assets\Scripts\Player\Player\PlayerState.cs" />
     <Compile Include="Assets\Scripts\Player\PlayerController.cs" />
-    <Compile Include="Assets\Scripts\SpeedUp.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="GenerateTargetFrameworkMonikerAttribute" />
